refactor: 降低代码耦合性
